Create merge_contigs_concoctclustering.py
The binning software CONCOCT suggests to split large contigs (>20 000 bp) into sub-contigs of 10 000 bp with a script that they provide with the program to give long contigs more weight. After binning and before creating fasta files for each bin, I prefer to merge these sub-contigs again. This script assigns the contigs that were split for binning to a bin based on the placement of the majority of the respective sub-contigs. It was written for SPAdes and IDBA-UD assemblies and may need adjustments for other contig naming systems.